Rename `create_attachment` -> `attachments`
